backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2